fix if login provide tables is empty error. change filed flagrow telegram id to bigint.

+++ migrations/2018_02_08_000000_add_telegram_columns_to_users_table.php
@@ -6,7 +6,7 @@ use Illuminate\Database\Schema\Builder;
 return [
     'up' => function (Builder $schema) {
         $schema->table('users', function (Blueprint $table) {
-            $table->unsignedInteger('flagrow_telegram_id')->nullable()->unique();
+            $table->bigInteger('flagrow_telegram_id')->nullable()->unique();
             $table->string('flagrow_telegram_error', 50)->nullable();
         });
     },
